+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Improve declaration with new syntax `@[a, b], c: [1, 2], 3;`, see [FEATURE#50](https://github.com/anole-lang/anole/issues/50)
+
 ### Changed
 
 - `return val1, val2;` equals `return [val1, val2]` now